+++ core/src/test/java/arez/ObserverTest.java
@@ -63,24 +63,6 @@ public class ObserverTest
     assertEquals( observer.getPriority(), Priority.NORMAL );
     assertEquals( observer.canObserveLowerPriorityDependencies(), false );
 
-    // All the hooks start out null
-    assertEquals( observer.getOnActivate(), null );
-    assertEquals( observer.getOnDeactivate(), null );
-    assertEquals( observer.getOnStale(), null );
-
-    final Procedure onActivate = new NoopProcedure();
-    final Procedure onDeactivate = new NoopProcedure();
-    final Procedure onStale = new NoopProcedure();
-
-    // Ensure hooks can be modified
-    observer.setOnActivate( onActivate );
-    observer.setOnDeactivate( onDeactivate );
-    observer.setOnStale( onStale );
-
-    assertEquals( observer.getOnActivate(), onActivate );
-    assertEquals( observer.getOnDeactivate(), onDeactivate );
-    assertEquals( observer.getOnStale(), onStale );
-
     observer.invariantState();
 
     assertEquals( context.getTopLevelObservers().get( observer.getName() ), observer );
@@ -612,39 +594,22 @@ public class ObserverTest
 
     final Observer observer = newReadOnlyObserver();
 
-    final TestProcedure onActivate = new TestProcedure();
-    final TestProcedure onDeactivate = new TestProcedure();
-    final TestProcedure onStale = new TestProcedure();
-
-    observer.setOnActivate( onActivate );
-    observer.setOnDeactivate( onDeactivate );
-    observer.setOnStale( onStale );
-
     assertEquals( observer.getState(), ObserverState.INACTIVE );
 
     observer.setState( ObserverState.INACTIVE );
 
     assertEquals( observer.getState(), ObserverState.INACTIVE );
     assertEquals( observer.isScheduled(), false );
-    assertEquals( onActivate.getCalls(), 0 );
-    assertEquals( onDeactivate.getCalls(), 0 );
-    assertEquals( onStale.getCalls(), 0 );
 
     observer.setState( ObserverState.UP_TO_DATE );
 
     assertEquals( observer.getState(), ObserverState.UP_TO_DATE );
     assertEquals( observer.isScheduled(), false );
-    assertEquals( onActivate.getCalls(), 1 );
-    assertEquals( onDeactivate.getCalls(), 0 );
-    assertEquals( onStale.getCalls(), 0 );
 
     observer.setState( ObserverState.POSSIBLY_STALE );
 
     assertEquals( observer.getState(), ObserverState.POSSIBLY_STALE );
     assertEquals( observer.isScheduled(), false );
-    assertEquals( onActivate.getCalls(), 1 );
-    assertEquals( onDeactivate.getCalls(), 0 );
-    assertEquals( onStale.getCalls(), 0 );
 
     observer.clearScheduledFlag();
     Arez.context().getScheduler().getPendingObservers().truncate( 0 );
@@ -653,24 +618,15 @@ public class ObserverTest
     observer.setState( ObserverState.STALE );
     assertEquals( observer.getState(), ObserverState.STALE );
     assertEquals( observer.isScheduled(), true );
-    assertEquals( onActivate.getCalls(), 1 );
-    assertEquals( onDeactivate.getCalls(), 0 );
-    assertEquals( onStale.getCalls(), 1 );
 
     observer.setState( ObserverState.UP_TO_DATE );
 
     assertEquals( observer.getState(), ObserverState.UP_TO_DATE );
-    assertEquals( onActivate.getCalls(), 1 );
-    assertEquals( onDeactivate.getCalls(), 0 );
-    assertEquals( onStale.getCalls(), 1 );
 
     observer.setState( ObserverState.STALE );
 
     assertEquals( observer.getState(), ObserverState.STALE );
     assertEquals( observer.isScheduled(), true );
-    assertEquals( onActivate.getCalls(), 1 );
-    assertEquals( onDeactivate.getCalls(), 0 );
-    assertEquals( onStale.getCalls(), 2 );
 
     final Observable<?> observable1 = newObservable();
     final Observable<?> observable2 = newObservable();
@@ -687,9 +643,6 @@ public class ObserverTest
     observer.setState( ObserverState.INACTIVE );
 
     assertEquals( observer.getState(), ObserverState.INACTIVE );
-    assertEquals( onActivate.getCalls(), 1 );
-    assertEquals( onDeactivate.getCalls(), 1 );
-    assertEquals( onStale.getCalls(), 2 );
 
     assertEquals( observer.getDependencies().size(), 0 );
     assertEquals( observable1.getObservers().size(), 0 );
@@ -702,15 +655,7 @@ public class ObserverTest
   {
     setupReadWriteTransaction();
 
-    final Observer observer = newReadOnlyObserver();
-
-    final TestProcedure onActivate = new TestProcedure();
-    final TestProcedure onDeactivate = new TestProcedure();
-    final TestProcedure onStale = new TestProcedure();
-
-    observer.setOnActivate( onActivate );
-    observer.setOnDeactivate( onDeactivate );
-    observer.setOnStale( onStale );
+    final Observer observer = newDerivation();
 
     assertEquals( observer.getState(), ObserverState.INACTIVE );
 
@@ -718,25 +663,16 @@ public class ObserverTest
 
     assertEquals( observer.getState(), ObserverState.INACTIVE );
     assertEquals( observer.isScheduled(), false );
-    assertEquals( onActivate.getCalls(), 0 );
-    assertEquals( onDeactivate.getCalls(), 0 );
-    assertEquals( onStale.getCalls(), 0 );
 
     observer.setState( ObserverState.UP_TO_DATE, false );
 
     assertEquals( observer.getState(), ObserverState.UP_TO_DATE );
     assertEquals( observer.isScheduled(), false );
-    assertEquals( onActivate.getCalls(), 1 );
-    assertEquals( onDeactivate.getCalls(), 0 );
-    assertEquals( onStale.getCalls(), 0 );
 
     observer.setState( ObserverState.POSSIBLY_STALE, false );
 
     assertEquals( observer.getState(), ObserverState.POSSIBLY_STALE );
     assertEquals( observer.isScheduled(), false );
-    assertEquals( onActivate.getCalls(), 1 );
-    assertEquals( onDeactivate.getCalls(), 0 );
-    assertEquals( onStale.getCalls(), 0 );
 
     observer.clearScheduledFlag();
     Arez.context().getScheduler().getPendingObservers().truncate( 0 );
@@ -745,24 +681,15 @@ public class ObserverTest
     observer.setState( ObserverState.STALE, false );
     assertEquals( observer.getState(), ObserverState.STALE );
     assertEquals( observer.isScheduled(), false );
-    assertEquals( onActivate.getCalls(), 1 );
-    assertEquals( onDeactivate.getCalls(), 0 );
-    assertEquals( onStale.getCalls(), 1 );
 
     observer.setState( ObserverState.UP_TO_DATE, false );
 
     assertEquals( observer.getState(), ObserverState.UP_TO_DATE );
-    assertEquals( onActivate.getCalls(), 1 );
-    assertEquals( onDeactivate.getCalls(), 0 );
-    assertEquals( onStale.getCalls(), 1 );
 
     observer.setState( ObserverState.STALE, false );
 
     assertEquals( observer.getState(), ObserverState.STALE );
     assertEquals( observer.isScheduled(), false );
-    assertEquals( onActivate.getCalls(), 1 );
-    assertEquals( onDeactivate.getCalls(), 0 );
-    assertEquals( onStale.getCalls(), 2 );
 
     final Observable<?> observable1 = newObservable();
     final Observable<?> observable2 = newObservable();
@@ -779,9 +706,6 @@ public class ObserverTest
     observer.setState( ObserverState.INACTIVE, false );
 
     assertEquals( observer.getState(), ObserverState.INACTIVE );
-    assertEquals( onActivate.getCalls(), 1 );
-    assertEquals( onDeactivate.getCalls(), 1 );
-    assertEquals( onStale.getCalls(), 2 );
 
     assertEquals( observer.getDependencies().size(), 0 );
     assertEquals( observable1.getObservers().size(), 0 );
@@ -805,9 +729,9 @@ public class ObserverTest
     final TestProcedure onDeactivate = new TestProcedure();
     final TestProcedure onStale = new TestProcedure();
 
-    observer.setOnActivate( onActivate );
-    observer.setOnDeactivate( onDeactivate );
-    observer.setOnStale( onStale );
+    computedValue.setOnActivate( onActivate );
+    computedValue.setOnDeactivate( onDeactivate );
+    computedValue.setOnStale( onStale );
 
     assertEquals( observer.getState(), ObserverState.INACTIVE );