+++ CHANGELOG.md
@@ -85,6 +85,9 @@
   set on the object if the hooks are non-null. In some scenarios this decreased runtime memory usage with the
   potential for a slight increase in code size in seom applications. It should be noted that In J2CL, this change
   actually resulted in a code size decrease if the application does not make use of the hooks.
+* **\[core\]** Move the `OnStale`, `OnActivate` and `OnDeactivate` hook methods from `arez.Observer` to
+  `arez.ComputedValue` to decrease memory pressure at runtime as the hooks are not required except for computed
+  properties.
 
 ### [v0.104](https://github.com/arez/arez/tree/v0.104) (2018-08-01)
 [Full Changelog](https://github.com/arez/arez/compare/v0.103...v0.104)